+++ CHANGELOG.md
@@ -17,3 +17,9 @@
 * Retain subcomponent graphs  
   [Sebastian Shanus](https://github.com/sebastianv1)
   [#42](https://github.com/square/Cleanse/issues/42)
+
+#### Improvements
+
+* Replace demo app's ErrorOptional type with Swift Foundation Result type
+  [Sihao Lu](https://github.com/DJBen)
+  [#126](https://github.com/square/Cleanse/pull/126)